# Improve Stoplight initialization performance by optimizing config merging