This change improves Stoplight initialization performance by approximately 12% by optimizing how configuration settings are merged during object creation.

The performance gain is achived by computing default settings once during `ConfigProvider` initialization instead of merging them on every light creation

Before:
```
Create Before Refactoring: 270,049.2 i/s (3.70 μs/i)
```
After:
```
Create After Refactoring: 301,795.9 i/s (3.31 μs/i)
```

Performance improvement: 12% faster initialization

#285